card layout & interaction improvements
