+++ src/components/structure/visualizing/artefact-card/artefact-card.scss
@@ -6,23 +6,21 @@
   flex-direction: column;
   background-color: $darker;
   position: relative;
-  transition: transform $tr-fast;
+  transition: opacity $tr-fast;
   overflow: hidden;
-  $self: &;
+  opacity: 0.9;
 
-  &__image {
-    padding: 1px;
-  }
+  $self: &;
 
   &__content {
     position: relative;
     flex-grow: 1;
-    //padding: $s;
     padding-top: $s;
     padding-bottom: $s;
-    border-top: solid $border-stroke-weight-s $accent;
+    border-top: solid $border-stroke-weight-s transparent;
     background-color: $darkest;
     color: $lighter;
+    transition: border-color $tr-slow;
   }
 
   &__title {
@@ -75,13 +73,15 @@
   }
 
   &:hover{
-    transform: scale(1.05);
-    box-shadow: $shadow-m;
-    z-index: $level-overall-nav;
+    opacity: 1;
 
     #{$self}__favorite{
       opacity: 1 !important;
       top: $xs !important;
+    }
+
+    #{$self}__content{
+      border-color: $accent;
     }
   }
 }
